# Conflicts:
#	TTMS/src/main/java/com/ttms/Mapper/SysDepartmentMapper.java
#	TTMS/src/main/java/com/ttms/service/ProductManage/ServiceImpl/ProjectService.java
#	TTMS/src/main/java/com/ttms/service/SystemManage/IGroupService.java

+++ TTMS/src/main/java/com/ttms/utils/BaseMapper.java
@@ -4,5 +4,5 @@ import tk.mybatis.mapper.additional.idlist.IdListMapper;
 import tk.mybatis.mapper.additional.insert.InsertListMapper;
 import tk.mybatis.mapper.common.Mapper;
 
-public interface BaseMapper<T> extends Mapper<T>, IdListMapper<T,Long>, InsertListMapper {
+public interface BaseMapper<T> extends Mapper<T>, IdListMapper<T,Integer>, InsertListMapper {
 }
